[refactor] 프로젝트 구조 변경 (#83)
* #82 style : ItemResponse와 ItemResponses 위치 변경

* #82 refactor : ItemService와 ItemQueryService를 하나의 ItemService로 관리하게끔 변경

클래스단에 붙어있는 @Transactional(readOnly = true)로 메서드에@Transactional(readOnly = true) 붙여져있는 어노테이션 삭제

* #82 refactor : ItemQueryService로 오류나는 부분 수정
