+++ backend/src/main/java/codesquard/app/api/sales/SalesItemController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import codesquard.app.api.item.response.ItemResponses;
 import codesquard.app.api.response.ApiResponse;
-import codesquard.app.api.response.ItemResponses;
 import codesquard.app.domain.sales.SalesStatus;
 import lombok.RequiredArgsConstructor;
