+++ backend/src/main/java/codesquard/app/api/wishitem/WishItemController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import codesquard.app.api.item.response.ItemResponses;
 import codesquard.app.api.response.ApiResponse;
-import codesquard.app.api.response.ItemResponses;
 import codesquard.app.domain.oauth.support.AuthPrincipal;
 import codesquard.app.domain.oauth.support.Principal;
 import codesquard.app.domain.wish.WishStatus;
