+++ backend/src/main/java/codesquard/app/api/wishitem/WishItemService.java
@@ -10,8 +10,8 @@ import codesquard.app.api.errors.errorcode.ItemErrorCode;
 import codesquard.app.api.errors.errorcode.MemberErrorCode;
 import codesquard.app.api.errors.errorcode.WishErrorCode;
 import codesquard.app.api.errors.exception.RestApiException;
-import codesquard.app.api.response.ItemResponse;
-import codesquard.app.api.response.ItemResponses;
+import codesquard.app.api.item.response.ItemResponse;
+import codesquard.app.api.item.response.ItemResponses;
 import codesquard.app.domain.item.Item;
 import codesquard.app.domain.item.ItemRepository;
 import codesquard.app.domain.member.Member;
