+++ backend/src/test/java/codesquard/app/api/item/ItemQueryServiceTest.java
@@ -67,7 +67,7 @@ class ItemQueryServiceTest extends IntegrationTestSupport {
 		wishRepository.save(wish);
 
 		// when
-		ItemDetailResponse response = itemQueryService.findDetailItemBy(item.getId(), member.getId());
+		ItemDetailResponse response = itemService.findDetailItemBy(item.getId(), member.getId());
 		// then
 		SoftAssertions.assertSoftly(softAssertions -> {
 			softAssertions.assertThat(response)
@@ -89,7 +89,7 @@ class ItemQueryServiceTest extends IntegrationTestSupport {
 		Long itemId = 9999L;
 		// when
 		Throwable throwable = Assertions.catchThrowable(
-			() -> itemQueryService.findDetailItemBy(itemId, member.getId()));
+			() -> itemService.findDetailItemBy(itemId, member.getId()));
 		// then
 		Assertions.assertThat(throwable)
 			.isInstanceOf(RestApiException.class)
