+++ backend/src/test/java/codesquard/app/api/item/ItemServiceTest.java
@@ -29,9 +29,9 @@ import codesquard.app.api.category.CategoryTestSupport;
 import codesquard.app.api.image.ImageUploader;
 import codesquard.app.api.item.request.ItemModifyRequest;
 import codesquard.app.api.item.request.ItemRegisterRequest;
+import codesquard.app.api.item.response.ItemResponse;
+import codesquard.app.api.item.response.ItemResponses;
 import codesquard.app.api.oauth.OauthFixedFactory;
-import codesquard.app.api.response.ItemResponse;
-import codesquard.app.api.response.ItemResponses;
 import codesquard.app.domain.category.Category;
 import codesquard.app.domain.image.Image;
 import codesquard.app.domain.item.Item;
